Add default_mv_refresh_partition_num to set default partition_refresh_num
