+++ test/lib/sr_sql_lib.py
@@ -919,14 +919,15 @@ class StarrocksSQLApiLib(object):
             count += 1
         tools.assert_equal("FINISHED", status, "wait alter table finish error")
 
-    def wait_async_materialized_view_finish(self, mv_name, check_count=60):
+    def wait_async_materialized_view_finish(self, mv_name, min_success_num = 1, check_count=60):
         """
         wait async materialized view job finish and return status
         """
         status = ""
         show_sql = "SHOW MATERIALIZED VIEWS WHERE name='" + mv_name + "'"
         count = 0
-        while count < check_count:
+        success_num = 0
+        while count < check_count and success_num < min_success_num:
             res = self.execute_sql(show_sql, True)
             status = res["result"][-1][12]
             if status != "SUCCESS":
@@ -934,7 +935,7 @@ class StarrocksSQLApiLib(object):
             else:
                 # sleep another 5s to avoid FE's async action.
                 time.sleep(1)
-                break
+                success_num += 1
             count += 1
         tools.assert_equal("SUCCESS", status, "wait aysnc materialized view finish error")
